Huy has done tasks
- LoginByGoogle
- Send OTP
- Verify OTP
- ResetPassword
- GetAccountByEmail
- GetAccountById
- Update Account

+++ Backend/SWP391.EventFlowerExchange/SWP391.EventFlowerExchange.FA24/SWP391.EventFlowerExchange.Application/AccountService.cs
@@ -8,25 +8,108 @@ namespace SWP391.EventFlowerExchange.Application
     public class AccountService : IAccountService
     {
         private IAccountRepository _repo;
+        private readonly Microsoft.AspNetCore.Identity.RoleManager<IdentityRole> _roleManager;
+        private readonly Microsoft.AspNetCore.Identity.UserManager<Account> _userManager;
 
-        public AccountService(IAccountRepository repo)
+        public AccountService(IAccountRepository repo, UserManager<Account> userManager, RoleManager<IdentityRole> roleManager)
         {
             _repo = repo;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
+        //SignIn
+        public async Task<string> SignInFromAPIAsync(SignIn signIn)
+        {
+            return await _repo.SignInAsync(signIn);
+        }
+
+        public async Task<string> SignInEmailFromAPIAsync(SignIn signIn)
+        {
+            return await _repo.SignInEmailAsync(signIn);
+        }
+
+        //SignUp
+        public async Task<IdentityResult> SignUpBuyerFromAPIAsync(SignUpBuyer signUp)
+        {
+            return await _repo.SignUpBuyerAsync(signUp);
+        }
+        public async Task<IdentityResult> SignUpSellerFromAPIAsync(SignUpSeller signUp)
+        {
+            return await _repo.SignUpSellerAsync(signUp);
+        }
         public async Task<IdentityResult> CreateStaffAccountFromAPIAsync(SignUpStaff staff)
         {
             return await _repo.CreateStaffAccountAsync(staff);
         }
+        public async Task<IdentityResult> CreateShipperAccountFromAPIAsync(SignUpShipper shipper)
+        {
+            return await _repo.CreateShipperAccountAsync(shipper);
+        }
 
+        //SendOTP
+        public async Task<bool> SendOTPFromAPIAsync(string email)
+        {
+            return await _repo.SendOTPAsync(email);
+        }
+
+        public async Task<bool> VerifyOTPFromAPIAsync(string email, string otp)
+        {
+            return await _repo.VerifyOTPAsync(email, otp);
+        }
+
+        //Reset password
+        public async Task<bool> ResetPasswordFromAPIAsync(string email, string newPassword)
+        {
+            return await _repo.ResetPasswordAsync(email, newPassword);
+        }
+
+        //CRUD Account
+        public async Task<Account> GetUserByEmailFromAPIAsync(Account account)
+        {
+            return await _repo.GetUserByEmailAsync(account);
+        }
+
+        public async Task<Account> GetUserByIdFromAPIAsync(Account account)
+        {
+            return await _repo.GetUserByIdAsync(account);
+        }
+
+        public async Task<IdentityResult> UpdateAccountFromAPIAsync(Account account)
+        {
+            return await _repo.UpdateAccountAsync(account);
+        }
         public async Task<IdentityResult> DeleteAccountFromAPIAsync(Account account)
         {
-            return await _repo.DeleteAccountAsync(account);
+            var userRoles = await _userManager.GetRolesAsync(account);
+
+            foreach (var role in userRoles)
+            {
+                if (role.ToLower().Contains("buyer")
+                    || role.ToLower().Contains("seller"))
+                {
+                    await _repo.DeleteAccountAsync(account);
+                    return IdentityResult.Success;
+                }
+            }
+
+            return IdentityResult.Failed();
         }
 
         public async Task<IdentityResult> RemoveAccountFromAPIAsync(Account account)
         {
-            return await _repo.RemoveAccountAsync(account);
+            var userRoles = await _userManager.GetRolesAsync(account);
+
+            foreach (var role in userRoles)
+            {
+                if (role.ToLower().Contains("admin")
+                    || role.ToLower().Contains("shipper"))
+                {
+                    await _repo.RemoveAccountAsync(account);
+                    return IdentityResult.Success;
+                }
+            }
+            return IdentityResult.Failed();
         }
 
         public async Task<List<Account>> SearchAccountsByAddressFromAPIAsync(string address)
@@ -38,17 +121,6 @@ namespace SWP391.EventFlowerExchange.Application
         {
             return await _repo.SearchAccountsBySalaryAsync(minSalary, maxSalary);
         }
-
-        public async Task<string> SignInFromAPIAsync(SignIn signIn)
-        {
-            return await _repo.SignInAsync(signIn);
-        }
-
-        public async Task<IdentityResult> SignUpBuyerFromAPIAsync(SignUp signUp)
-        {
-            return await _repo.SignUpBuyerAsync(signUp);
-        }
-
         public async Task<List<Account>> ViewAllAccountFromAPIAsync()
         {
             return await _repo.ViewAllAccountAsync();
@@ -59,24 +131,9 @@ namespace SWP391.EventFlowerExchange.Application
             return await _repo.ViewAllAccountByRoleAsync(role);
         }
 
-        public async Task<IdentityResult> SignUpSellerFromAPIAsync(SignUpSeller signUp)
-        {
-            return await _repo.SignUpSellerAsync(signUp);
-        }
-
-        public async Task<IdentityResult> CreateShipperAccountFromAPIAsync(SignUpShipper shipper)
-        {
-            return await _repo.CreateShipperAccountAsync(shipper);
-        }
-
         public async Task<List<Account>> SearchShipperByAddressFromAPIAsync(string address)
         {
             return await _repo.SearchShipperByAddressAsync(address);
-        }
-
-        public async Task<Account> GetUserByIdFromAPIAsync(Account account)
-        {
-            return await _repo.GetUserByIdAsync(account);
         }
     }
 }
